@@ -1,65 +1,25 @@
-import { Routes, Route, useLocation } from 'react-router-dom';
-import { Home } from './pages/Home';
-import { RequireAuth } from './RequireAuth';
-import { PageOne } from './pages/Page1';
-import { PageTwo } from './pages/Page2';
-import { SignUp } from './pages/SignUp';
-import { SignIn } from './pages/SignIn';
-import { useAuth } from './context/use-auth';
-import { Layout } from './components/Layout';
-import { NoMatch } from './pages/NoMatch';
-import { useEffect } from 'react';
-import { useNavigate } from 'react-router';
-import { LocationState } from './type';
+import React from 'react';
+import logo from './logo.svg';
+import './App.css';
 
-const App = () => {
-
-  const { isError, isLoading, refreshToken } = useAuth();
-  let location = useLocation();
-  let navigate = useNavigate()
-  let from = (location.state as LocationState)?.from?.pathname || '/';
-
-  const attemptTokenRefresh = async () => {
-    await refreshToken();
-    if (!isError) navigate(from, { replace: true })
-  }
-
-  useEffect(() => {
-    attemptTokenRefresh();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
-  const refreshingAnimation = (
-    <main className="main">
-      <h1>Please Wait.....</h1>
-    </main>
-  )
-
-  if (isLoading) return refreshingAnimation
-
+function App() {
   return (
-    <Routes>
-      <Route path="/" element={<Layout />}>
-        <Route index element={<Home />} />
-        <Route path="/signin" element={<SignIn />} />
-        <Route path="/signup" element={<SignUp />} />
-        <Route path="/page-one" element={
-          <RequireAuth>
-            <PageOne />
-          </RequireAuth>
-        }
-        />
-        <Route
-          path="/page-two"
-          element={
-            <RequireAuth>
-              <PageTwo />
-            </RequireAuth>
-          }
-        />
-        <Route path="*" element={<NoMatch />} />
-      </Route>
-    </Routes>
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.tsx</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header>
+    </div>
   );
 }
 
